updated subscription feature, added link to custom dashboard in nav, updated nav styling
