+++ frontend/src/constants/assets.js
@@ -34,6 +34,6 @@ export { default as CLOSE } from "../assets/images/close.svg";
 export { default as TEAM } from "../assets/images/Team.svg";
 export { default as HOME } from "../assets/images/home.svg";
 export { default as COMPANY } from "../assets/images/Company.svg";
-export { default as BELL } from "../assets/images/bell.svg";
+export { default as SUBSCRIBE } from "../assets/images/subscribe.svg";
 export { default as UNSUBSCRIBE } from "../assets/images/Notifications.svg";
 export { default as YELLOW_HAZARD } from "../assets/images/Hazard.svg";
