+++ frontend/src/styles/generic/layout.scss
@@ -268,6 +268,7 @@ body {
 
 .img-lg {
   height: 30px;
+  font-size: 26px;
 }
 
 .vertical-align {
